+++ recipes/compute_slurm_config.rb
@@ -35,16 +35,8 @@ mount '/opt/slurm' do
   retry_delay 6
 end
 
-# Check to see if there is GPU on the instance, only execute run_nvidiasmi if there is GPU and nvidia installed
-if graphic_instance?
-  if nvidia_installed?
-    execute "run_nvidiasmi" do
-      command 'nvidia-smi'
-    end
-  else
-    Chef::Log.warn("GPU instance but no Nvidia drivers found")
-  end
-end
+# Check to see if is GPU instance with Nvidia installed
+Chef::Log.warn("GPU instance but no Nvidia drivers found") if graphic_instance? && !nvidia_installed?
 
 cookbook_file '/etc/systemd/system/slurmd.service' do
   source 'slurmd.service'